requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.34.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
